Rehash the native internals.
The first implementation of this didn't really resemble the patch
suggested to RJBS for Dzil.

It is now more conformant to that, and isn't sneakily automatically
turning itself on if there just so happens to be a munge method
available.

Now you have to specify you want to use the api, and specify which api
you're using.